Implement token authorization
We made this change to offer API endpoints with access control in the future.

Only users in the database should be able to access non-public endpoints. The frontend can generate access tokens for users with an active account using the endpoint /auth/get_access_token/.

You can define the validity period of these tokens in the .env file.